+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.5.0 - Unreleased
 - Changed keybinding for wireframe mode from W to T
 - Added keybindings to effect the orientation of the rocket
+- Added keybindings to effect the throttle of the rocket
 
 # v0.4.0 - 2019-08-24
 - Clamp camera altitude to [1 mm, 1 Tm]